Singularize file_path when injecting into class
Railties assumes that `file_name`, set in
[`assign_names!`](https://github.com/rails/rails/blob/4a446220a67e115c04a2b562dce92981a8e18c78/railties/lib/rails/generators/named_base.rb#L177-L181)
as a derivative of `name` which can be plural or singular, will always
be singular when it implicitly aliases it in
[`singular_name`](https://github.com/rails/rails/blob/4a446220a67e115c04a2b562dce92981a8e18c78/railties/lib/rails/generators/named_base.rb#L36-L38).

This is really a rails/rails issue, but to resolve #148 for ourselves
we'll call `singularize` on `file_path` which inherits this assumption.
It seems to me that the solution really would be to singularize `name`
in NamedBase before calling `assign_names` since every instance of
`plural_*` calls pluralize.